Fix incorrect regular expression literal grammar definition
This changeset fixes the previously incorrect definition of the <regexp>
grammar production. Previously, this production only included syntax for
regular expression literals of the form

        r`regexp`

which was the first form of regular expression literals added to the
language during the pre-public Mellifera R&D phase. At the time this
first form was added, regular expression types and values were given the
name "regex" rather than the name "regexp" used today. With this
changeset, the <regex> production name has been changed to <regexp>, and
the two forms of regular expression literals

        r"regexp"
        r`regexp`

have been properly defined in the language grammar file.