Replace math domain error propagation with special case return values
This changeset replaces the "math domain error" errors produced by
builtin math functions with special case return values for each builtin.
Previously, when a ValueError was raised from a Python math.* function
corresponding to certain Mellifera math::* functions, that exception was
translated into a Mellifera error verbatim. This closely tied the
implementation behavior of the Mellifera language to the implementation
behavior of the Python interpreter, and led to some questionable results
that did not necessarily align with user expectations. This changeset
adds additional input and error checking to handle special cases for
these math functions.

Special care was taken to align the behavior of the builtin math
functions with the behavior of the Go math package. This was done both
because the Go math package has reasonable handling of special cases
that align with user expectations, and because there are plans to
eventually port the Mellifera interpreter to Go, in which case greater
congruence between math behavior is a desired property.