Revamp readme with a practical language example
This changeset updates the project readme to provide a better language
overview for new readers. I was unhappy with the structure and flow of
the previous readme, which failed to explain the philosophy and goals of
the language and did not show any "real-world" code. This changeset
introduces a new example program, examples/word-count.mf, intended to
show some of the syntax and semantics of Mellifera using a practical
program. The updated readme places this new example right at the top of
the file, giving readers an immediate language showcase when looking at
the repository for the first time. Additional changes were made to the
readme to accompany other example programs with brief text explanations
on how those examples relate to the core design and goals of the
language.

During the development of the word-count.mf program, the following
builtins were added to the language as nice-to-have quality-of-life
improvements:

+ regexp::init
+ regexp::split
+ regexp::sub
+ map::keys
+ map::values
+ map::pairs
+ re::split
+ re::sub

Additionally, the Python Mellifera implementation's use of Google's re2
library has had a bug(?) fixed where calling into the C++ implementation
of re2.compile with an invalid regex no longer dumps an Abseil warning
to stderr.