Add json::encode_ex
This changeset introduces json::encode_ex, a variation on json::encode
that allows for more precise control over the JSON encoding process.
Currently, json::encode_ex accepts one encoding option, "indent", which
must be either an integer representing the number of spaces to indent,
or a string used for each level of indent.