Optimize the performance of Environment lookup and assignment operations
This changeset updates the Environment class to use a dict to directly
map String identifiers to their corresponding values within a lexical
scope. Previously, an environment contained a Map data store that
allowed for arbitrary keys in the environment mapping. This was
originally introduced in Mellifera's predecessor language with the idea
that the environment would be a map that could be manipulated like any
other lvalue in the language. However, this vision never materialized in
practice, and there is no reason to pay a performance cost for that
decision in the present.

This optimization opportunity was discovered when profiling programs in
which a single tight loop or recursive call dominated the execution time
of the program. In these scenarios, environment creation, lookup, and
assignment accounted for a disproportionately large percentage of the
execution time compared to other operations within the same tight loops
and recursive calls. This changeset aims to reduce the total amount of
work that needs to be performed by environment operations in these
scenarios by using simpler data structures which attempt to keep more of
the execution time within the native code of CPython rather than in the
mf.py Python code that is already being interpreted.

Prior to this changeset, average execution of the recursive Fibonacci
example would look something like:

    $ time python3 mf.py examples/fibonacci-recursive.mf 25
    75025

    real    0m4.628s
    user    0m2.276s
    sys     0m2.865s

    $ time mf examples/fibonacci-recursive.mf 25
    75025

    real    0m2.810s
    user    0m3.099s
    sys     0m0.023s

After this changeset, average execution of the recursive Fibonacci
example would look something like:

    $ time python3 mf.py examples/fibonacci-recursive.mf 25
    75025

    real    0m1.934s
    user    0m1.199s
    sys     0m0.944s

    $ time mf examples/fibonacci-recursive.mf 25
    75025

    real    0m1.617s
    user    0m1.789s
    sys     0m0.000s

Although these results look spectacular on paper, it should be noted
that this 40-60% performance improvement was specifically for a case in
which the performance was heavily tied to environment operations to
begin with, and the magnitude of this performance improvement will not
be this large for all Mellifera programs.

Additional effort was put into investigating whether directly using the
bytes data of strings would yield additional performance improvements.
However, without any sort of formal performance framework set up, it was
difficult to discern whether that additional change would have had a
statistically significant effect on program runtime.

This commit also removes the baseenv builtin function. This function was
not used anywhere, had no tests, and should be considered a vestigal
structure from the earlier environment design.