Rename the HIVE data format and encoding to COMB
After some discussion, using the name COMB instead of HIVE makes more
sense for a data format, as "honeycomb" elicits imagery of an actual
physical structure where "beehive" is more of a concept. Either way the
name is arbitrary, but after having it sit on the tongue for a day,
saying things like "this comb file" and "this is comb-encoded" felt a
bit better than "this hive file" and "this is hive-encoded".