Avoid creating new objects when copying immutable values
This changeset adds an optimization that improves performance when
copying immutable primitive values by returning a reference to the
original object being copied from instead of created a new object for
every copy operation.

It has been possible to implement this optimization since the start of
the project, but this optimization was avoided until now as it was
aesthetically pleasing when the operation `a = b` produced different
objects with different addresses for `a` and `b` following the
conceptual idea that `a` and `b` are wholly separate entities. However,
with the recent focus on performance work in the reference interpreter,
and with architectural design being solidified in preparation for a port
of the reference interpreter from Python to Go, it seems appropriate to
lock this optimization in now.

Performance testing for this changeset was carried out on a Windows
machine under WSL with other services running in the background, so
reported metrics should be taken with a grain of salt due to variance in
system load and general system performance differences between WSL and
native Linux/macOS builds of the compiled interpreter.

Prior to this changeset, timing of the recursive Fibonacci example and
word count example would look something like:

    $ for _ in $(seq 3); do /usr/bin/time -f '%E' python3 mf.py examples/fibonacci-recursive.mf 25 >/dev/null; done
    0:01.95
    0:01.91
    0:01.91
    $ for _ in $(seq 3); do /usr/bin/time -f '%E' mf examples/fibonacci-recursive.mf 25 >/dev/null; done
    0:01.51
    0:01.45
    0:01.45

    $ for _ in $(seq 3); do /usr/bin/time -f '%E' python3 mf.py examples/word-count.mf gutenberg-71.txt >/dev/null; done
    0:01.70
    0:01.68
    0:01.68
    $ for _ in $(seq 3); do /usr/bin/time -f '%E' mf examples/word-count.mf gutenberg-71.txt >/dev/null; done
    0:02.52
    0:02.51
    0:02.54

After this changeset, timing of the recursive Fibonacci example and word
count example would look something like:

    $ for _ in $(seq 3); do /usr/bin/time -f '%E' python3 mf.py examples/fibonacci-recursive.mf 25 >/dev/null; done
    0:01.73
    0:01.81
    0:01.87
    $ for _ in $(seq 3); do /usr/bin/time -f '%E' mf examples/fibonacci-recursive.mf 25 >/dev/null; done
    0:01.25
    0:01.26
    0:01.26

    $ for _ in $(seq 3); do /usr/bin/time -f '%E' python3 mf.py examples/word-count.mf gutenberg-71.txt >/dev/null; done
    0:01.66
    0:01.62
    0:01.62
    $ for _ in $(seq 3); do /usr/bin/time -f '%E' mf examples/word-count.mf gutenberg-71.txt >/dev/null; done
    0:02.36
    0:02.37
    0:02.33