Add support for ad-hoc command execution with -c / --command
This changeset introduces new flags, -c and --command, that allow users
to execute Mellifera code from a string provided as an argument to the
interpreter. The changeset also replaces Python's automated argument
parsing from the argparse module with custom argument parsing in an
attempt to make interpreter argument parsing more language agnostic in
preparation for a port of the reference interpreter from Python to Go.

    $ mf -c 'println("hello world");'
    hello world
    $ mf --command 'println("hello world");'
    hello world
    $ mf --command='println("hello world");'
    hello world
    $ mf --command='println(module::file);'
    mf.py
    $ mf examples/program-arguments.mf foo bar baz
    program arguments: ["examples/program-arguments.mf", "foo", "bar", "baz"]
    $ mf examples/program-arguments.mf -- -c -h
    program arguments: ["examples/program-arguments.mf", "-c", "-h"]
    $ mf -- examples/program-arguments.mf -c -h
    program arguments: ["examples/program-arguments.mf", "-c", "-h"]
    $ mf <<EOF
    println("hello world");
    EOF
    >>> hello world
    >>>