Use a single null object for all null values
This changeset is a follow-up to earlier changes focused on reducing the
number of objects created when copying immutable values. The null value
in Mellifera is a stateless object with no metamap, and is thus a prime
candidate for optimization with a singleton. This changeset has the
effect that all null values now share the same reference address.