Rename vector::iterator to vector::into_iterator
Following along with the theme of making conversions from TypeA to TypeB
using the metafunction TypeA::into_TypeB, this changeset changes
vector::iterator to vector::into_iterator. With this changeset, the
vector -> iterator conversion is done by somevector::into_iterator and
the reverse is done with someiterator::into_vector, providing a more
intuitive symmetry for the vector conversions.