Add mf.go
This is currently just a placeholder that prints a hello world message.
The Mellifera library code will remain in mellifera.go, and the actual
interpreter executable will live in cmd/mf/mf.go. The plan is to slowly
bring mf.go up to parity with mf.py over time as the lexer, parser, and
tree-walk evaluator are slowly added to the Go implementation.