Add initial parsing and evaluation support to the Go implementation
This changeset lays the groundwork for the parsing and evaluation of the
abstract syntax tree within the Go implementation of Mellifera. The
changeset introduces new Parser and Environment types, as well as new
AST types and associated interfaces for expression and statement nodes.

This is a minimal working changeset, so the only programs that can
successfully be parsed and evaluated by this implementation are those
which consists of expression statements containing null expressions.

    $ mf --command 'null;'
    $ mf --command 'null'
    [<command>:1] expected `;`, found `end-of-file`
    $ mf --command 'null foo'
    [<command>:1] expected `;`, found `identifier`
    $ mf --command 'null
    > foo'
    [<command>:2] expected `;`, found `identifier`
    $ mf --dump-tokens --command 'null;'
    [
        {
            "kind": "null",
            "literal": "null",
            "location": {
                "file": "<command>",
                "line": 1
            }
        },
        {
            "kind": ";",
            "literal": ";",
            "location": {
                "file": "<command>",
                "line": 1
            }
        }
    ]