Add --dump-ast flag
This changeset introduces a new flag to mf.py and mf.go, --dump-ast,
which will print a comb-encoded representation of the abstract syntax
tree parsed from a command or source file to stdout.

This flag will be used to test the compatibility of the Python and Go
implementations of Mellifera at the abstract syntax tree level. At the
time of writing, the Python implementation only implements
AST-to-Mellifera value conversions for the subset of abstract syntax
tree nodes present within both versions of the interpreter.

    $ python3 mf.py --dump-ast --command 'null;'
    {
        "kind": "AstProgram",
        "location": {
            "file": "<command>",
            "line": 1
        },
        "statements": [
            {
                "kind": "AstStatementExpression",
                "location": {
                    "file": "<command>",
                    "line": 1
                },
                "expression": {
                    "kind": "AstExpressionNull",
                    "location": {
                        "file": "<command>",
                        "line": 1
                    }
                }
            }
        ]
    }

    $ make build-go >/dev/null && ./bin/mf --dump-ast --command 'null;'
    {
        "kind": "AstProgram",
        "location": {
            "file": "<command>",
            "line": 1
        },
        "statements": [
            {
                "kind": "AstStatementExpression",
                "location": {
                    "file": "<command>",
                    "line": 1
                },
                "expression": {
                    "kind": "AstExpressionNull",
                    "location": {
                        "file": "<command>",
                        "line": 1
                    }
                }
            }
        ]
    }

This changeset also adds tools/validate-dump-compatibility.sh and
tools/validate-dump-compatibility.mf, a set of files intended to help
with the port of Mellifera from Python to Go. This script will
automatically check for differences in the output of --dump-tokens and
--dump-ast invocations of each respective interpreter on the test file
tools/validate-dump-compatibility.mf. As new functionality is added to
the Go implementation, this Mellifera script should be updated with some
use of that functionality to ensure compatibility between the
implementations without requiring all of the supporting functionality
required by the full suite of golden tests.