Use a single true and false object for all boolean values
This changeset adds singletons for the true and false boolean values.
This builds on some previous work where null was made into a singleton
in order to save on the total number of objects created when copying
immutable values. This changeset has the effect that all true boolean
values now share the same address, and all false boolean values now
share the same address.