Use string source instead of []rune source in the Go implementation's Lexer
Originally, the type []rune was used for the lexer source in order to
match the Python implementation's str source for rune-by-rune lexing.
However, the unicode/utf8 package contains the DecodeRuneInString
function, which allows one to easily decode individual runes at the
start of the source string, achieving similar behavior while still
giving us a string that can be used for regexp matching.